Its working loll

+++ job/job.go
@@ -5,7 +5,7 @@ import "distributed/structures"
 type Job struct {
 	Name       string             `json:"name"`
 	PointCount int                `json:"pointCount"`
-	Ration     int                `json:"p"`
+	Ration     float32            `json:"p"`
 	Width      int                `json:"width"`
 	Height     int                `json:"height"`
 	Points     []structures.Point `json:"-"`
